awscdk/memgraph/requirements.txt
-e .
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
configparser
boto3